Create new shopify util folder in blockchain-components
