+++ packages/blockchain-components/src/utils/shopify/const.ts
@@ -1,0 +1,8 @@
+export const shopifyServices = {
+  CART: {name: 'Cart', pathname: 'cart'},
+  CHECKOUT: {name: 'Checkout', pathname: 'checkouts'},
+  COLLECTION: {name: 'Collections', pathname: 'collections'},
+  OTHER: {name: 'Other', pathname: undefined},
+  PAGE: {name: 'Pages', pathname: 'pages'},
+  PDP: {name: 'Product Details Page', pathname: 'products'},
+};
